requirements.txt
Django==4.0.6
djangorestframework==3.13.1
psycopg2-binary==2.8.5